% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR_based.R
\name{APclassify}
\alias{APclassify}
\title{Classify genes based on types of APUs}
\usage{
APclassify(APtestObject, significanceThreshold = 0.05,
  promoterTest = "both", geneTest = "both")
}
\arguments{
\item{APtestObject}{APtest object.}

\item{significanceThreshold}{FDR-corrected p-value threshold for significance.}

\item{promoterTest}{Not yet implemented.}

\item{geneTest}{Not yet implemented.}
}
\value{
APclassify object
}
\description{
Classfiy multipromoter genes into 7 classes based on types of APU.
}
\examples{
# ADD EXAMPLES HERE
}

