% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR_based.R
\name{APfit}
\alias{APfit}
\title{Fit GLMs at the gene and promoter level}
\usage{
APfit(promoterCounts, modelMatrix, geneIds)
}
\arguments{
\item{promoterCounts}{matrix: Number of CAGE tags in each promoter.}

\item{modelMatrix}{matrix: Full rank model matrix describing the experiment.}

\item{geneIds}{character: Vector marking which genes promoters belong two.}
}
\value{
APfit object
}
\description{
Uses edgeR to fit a robust QL GLM to promoters and genes
}
\examples{
# ADD EXAMPLES HERE
}

