% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR_based.R
\name{APtest}
\alias{APtest}
\title{Perform test for APU}
\usage{
APtest(APfitObject, contrast)
}
\arguments{
\item{APfitObject}{APfit Object.}

\item{contrast}{numeric or matrix: A single contrast or a contrast matrix to be tested.}
}
\value{
RETURN APtest object
}
\description{
Use edgeR's diffspliceDGE to perform flat DE and nested APU analyses
}
\examples{
# ADD EXAMPLES HERE
}

