% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.APclassify}
\alias{plot.APclassify}
\title{Plot APU classification results}
\usage{
\method{plot}{APclassify}(o, contrast = NULL)
}
\arguments{
\item{o}{APclassify object.}

\item{contrast}{Name of contrast in APclassify object.}
}
\value{
RETURN DESCRIPTION
}
\description{
Plots either the result of a single contrast, or compares the output of multiple contrasts.
}
\examples{
# ADD EXAMPLES HERE
}

