% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.APfit}
\alias{plot.APfit}
\title{Plot basic statistics for a single gene}
\usage{
\method{plot}{APfit}(o, gene, stat, grouping = NULL)
}
\arguments{
\item{o}{APfit object.}

\item{gene}{character: gene name.}

\item{stat}{character: "expression" or "BCV".}

\item{grouping}{factor: groups to summarize expression by.}
}
\value{
ggplot2 object
}
\description{
Plots mean expression by a grouping of samples or the BCV from a GLM fit.
}
\examples{
# ADD EXAMPLES HERE
}

