% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.APtest}
\alias{plot.APtest}
\title{Plot DE statistics for a single gene}
\usage{
\method{plot}{APtest}(o, gene, stat)
}
\arguments{
\item{o}{APtests object}

\item{gene}{character: gene name}

\item{stat}{character: Either "logFC" or "pValue"}
}
\value{
ggplot2 object
}
\description{
Plot either logFCs or p-values for a single gene, at both gene and promoter levels.
}
\examples{
# ADD EXAMPLES HERE
}

