% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.APclassify}
\alias{print.APclassify}
\title{Print basic info about APclassify}
\usage{
\method{print}{APclassify}(o)
}
\arguments{
\item{o}{APClassify object}
}
\description{
Short representation fo APclassify object
}
\examples{
# ADD EXAMPLES HERE
}

