% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.APfit}
\alias{print.APfit}
\title{#' @export
print.APfit <- function(o){
	message("APfit object from the the APhunteR package")
}
Print basic info about APfit}
\usage{
\method{print}{APfit}(o)
}
\arguments{
\item{o}{APfit object}
}
\description{
Short representation fo APfit object
}
\examples{
# ADD EXAMPLES HERE
}

