% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.APtest}
\alias{print.APtest}
\title{#' @export
print.APtest <- function(o){
	message("APtest object from the the APhunteR package")
}
Print basic info about APtest}
\usage{
\method{print}{APtest}(o)
}
\arguments{
\item{o}{APtest object}
}
\description{
Short representation fo APtest object
}
\examples{
# ADD EXAMPLES HERE
}

