% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR_based.R
\name{singleClassify}
\alias{singleClassify}
\title{Internal function: Classify single contrast}
\usage{
singleClassify(con, p = 0.05, promoterTest = "both", geneTest = "both")
}
\arguments{
\item{con}{numeric: contrast}

\item{p}{numeric: FDR-corrected p-value threshold.}

\item{promoterTest}{See APclassify.}

\item{geneTest}{See APclassify.}
}
\value{
See APclassify
}
\description{
Called by APclassify
}
\examples{
# ADD EXAMPLES HERE
}

