% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR_based.R
\name{singleContrast}
\alias{singleContrast}
\title{Internal function: Test single contrast}
\usage{
singleContrast(o, contrast)
}
\arguments{
\item{o}{APfit object}

\item{contrast}{numeric: Valid contrast.}
}
\value{
RETURN list with info returned by APtest
}
\description{
Called by APtest
}
\examples{
# ADD EXAMPLES HERE
}

