% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacking.R
\name{stackLegends}
\alias{stackLegends}
\title{Internal function: Plot a ggplot2 legend over multiple colums}
\usage{
stackLegends(n)
}
\arguments{
\item{n}{integer: Number of columns.}
}
\value{
ggplot2 object
}
\description{
Called by stackPlots
}
\examples{
# ADD EXAMPLES HERE
}

