% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacking.R
\name{stackPlots}
\alias{stackPlots}
\title{Stack multiple single gene plots}
\usage{
stackPlots(ggplots, legendColumns = 2)
}
\arguments{
\item{ggplots}{list: list of ggplot2 objects.}

\item{legendColumns}{integer: number of columns in legend.}
}
\value{
None
}
\description{
This function will appropriately stack plots by reformating legends, removing x-axis labels and aligning plot windows.
}
\examples{
# ADD EXAMPLES HERE
}

